requirements.txt
streamlit==1.24.0
pandas==1.5.3
PyYAML==6.0
requests==2.28.1
